requirements.txt
pandas
psycopg2-binary
python-csv
numpy
alive_progress